backend/requirements.txt
amqp==5.3.1
asgiref==3.11.1
attrs==25.4.0
billiard==4.2.4
celery==5.3.4
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==6.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docker==6.1.3
drf-spectacular==0.27.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
packaging==26.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
redis==5.0.1
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
websocket-client==1.9.0
